@@ -28,31 +28,31 @@
     "vue": ">=3.4.0"
   },
   "dependencies": {
-    "@rollup/plugin-commonjs": "^25.0.7",
+    "@rollup/plugin-commonjs": "^26.0.1",
     "@rollup/plugin-node-resolve": "^15.2.3",
-    "@vitejs/plugin-vue": "^5.0.4",
-    "@vitejs/plugin-vue-jsx": "^3.1.0",
+    "@vitejs/plugin-vue": "^5.0.5",
+    "@vitejs/plugin-vue-jsx": "^4.0.0",
     "camelcase": "^8.0.0",
     "chalk": "^5.3.0",
     "consola": "^3.2.3",
-    "esbuild": "^0.21.3",
+    "esbuild": "^0.23.0",
     "fast-glob": "^3.3.2",
     "fs-extra": "^11.2.0",
-    "gulp": "^4.0.2",
+    "gulp": "^5.0.0",
     "gulp-autoprefixer": "^9.0.0",
     "gulp-clean-css": "^4.3.0",
     "gulp-filter": "^9.0.1",
     "gulp-rename": "^2.0.0",
     "gulp-sass": "^5.1.0",
-    "jsdom": "^24.0.0",
+    "jsdom": "^24.1.0",
     "parse-json": "^8.1.0",
-    "prettier": "^3.2.5",
+    "prettier": "^3.3.2",
     "rimraf": "^5.0.7",
-    "rollup": "^4.17.2",
+    "rollup": "^4.18.0",
     "rollup-plugin-esbuild": "^6.1.1",
-    "sass": "^1.77.2",
-    "ts-morph": "^22.0.0",
-    "tslib": "^2.6.2"
+    "sass": "^1.77.6",
+    "ts-morph": "^23.0.0",
+    "tslib": "^2.6.3"
   },
   "devDependencies": {
     "@rollup/plugin-typescript": "^11.1.6",
@@ -63,12 +63,12 @@
     "@types/gulp-filter": "^3.0.39",
     "@types/gulp-rename": "^2.0.6",
     "@types/gulp-sass": "^5.0.4",
-    "@types/jsdom": "^21.1.6",
+    "@types/jsdom": "^21.1.7",
     "@types/parse-json": "^4.0.2",
     "@types/prettier": "^2.7.3",
     "@types/rimraf": "^3.0.2",
-    "typescript": "^5.4.5",
-    "vite": "^5.2.11",
-    "vue": "^3.4.27"
+    "typescript": "^5.5.3",
+    "vite": "^5.3.3",
+    "vue": "^3.4.31"
   }
 }
