@@ -1,5 +1,6 @@
 {
   "name": "lc-library",
+  "type": "module",
   "private": true,
   "scripts": {
     "preinstall": "npx only-allow pnpm",
@@ -12,7 +13,7 @@
     "docs:build": "pnpm -C docs docs:build",
     "changelogs": "pnpm -r --parallel changelog && pnpm prettier",
     "prettier": "prettier --config .prettierrc --write ./**/*.md",
-    "lint": "eslint . --ext .js,.ts,.jsx,.tsx,.vue --fix",
+    "lint": "eslint",
     "vitest": "vitest",
     "vitest:c": "vitest run --coverage",
     "vitest:u": "vitest --coverage --ui",
@@ -34,45 +35,36 @@
     "pnpm": ">=8.0.0"
   },
   "devDependencies": {
-    "@changesets/cli": "^2.27.3",
+    "@changesets/cli": "^2.27.7",
     "@commitlint/cli": "^19.3.0",
     "@commitlint/config-conventional": "^19.2.2",
     "@esbuild-kit/cjs-loader": "^2.4.4",
-    "@noahyu/eslint-config-vue": "^1.3.0",
+    "@noahyu/conventional-changelog-base": "^1.0.1",
+    "@noahyu/eslint-config-vue": "^1.4.2",
     "@noahyu/lc-helpers": "workspace:*",
     "@noahyu/lc-ui": "workspace:*",
-    "@types/node": "20.12.12",
+    "@types/node": "20.14.9",
     "@types/rimraf": "^3.0.2",
-    "@typescript-eslint/eslint-plugin": "^7.9.0",
-    "@typescript-eslint/parser": "^7.9.0",
-    "@vitejs/plugin-vue": "^5.0.4",
-    "@vitejs/plugin-vue-jsx": "^3.1.0",
+    "@vitejs/plugin-vue": "^5.0.5",
+    "@vitejs/plugin-vue-jsx": "^4.0.0",
     "@vitest/coverage-v8": "^1.6.0",
     "@vitest/ui": "^1.6.0",
-    "@vue/compiler-dom": "3.4.27",
+    "@vue/compiler-dom": "3.4.31",
     "@vue/test-utils": "^2.4.6",
-    "c8": "^9.1.0",
+    "c8": "^10.1.2",
     "chalk": "^5.3.0",
     "commitizen": "^4.3.0",
-    "compare-func": "^2.0.0",
     "consola": "^3.2.3",
-    "conventional-changelog-angular": "^8.0.0",
-    "conventional-changelog-cli": "^5.0.0",
-    "cz-git": "^1.9.1",
-    "eslint": "^9.3.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-prettier": "^5.1.3",
-    "eslint-plugin-vue": "^9.26.0",
+    "cz-git": "^1.9.3",
     "fs-extra": "^11.2.0",
     "husky": "^9.0.11",
-    "jsdom": "^24.0.0",
-    "lint-staged": "^15.2.2",
-    "prettier": "^3.2.5",
+    "jsdom": "^24.1.0",
+    "lint-staged": "^15.2.7",
     "rimraf": "^5.0.7",
-    "typescript": "^5.4.5",
-    "vite": "^5.2.11",
+    "typescript": "^5.5.3",
+    "vite": "^5.3.3",
     "vitest": "^1.6.0",
-    "vue": "^3.4.27"
+    "vue": "^3.4.31"
   },
   "lint-staged": {
     "*.{js,ts,jsx,tsx,vue}": "eslint --fix"
@@ -81,5 +73,6 @@
     "commitizen": {
       "path": "./node_modules/cz-git"
     }
-  }
+  },
+  "packageManager": "pnpm@9.4.0+sha512.f549b8a52c9d2b8536762f99c0722205efc5af913e77835dbccc3b0b0b2ca9e7dc8022b78062c17291c48e88749c70ce88eb5a74f1fa8c4bf5e18bb46c8bd83a"
 }
