@@ -1,6 +1,7 @@
 {
   "name": "@noahyu/lc-helpers",
   "version": "1.5.0",
+  "type": "module",
   "description": "LC Helpers",
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
@@ -36,7 +37,7 @@
   },
   "dependencies": {},
   "devDependencies": {
-    "typescript": "^5.4.5",
-    "vue": "^3.4.27"
+    "typescript": "^5.5.3",
+    "vue": "^3.4.31"
   }
 }
